Create seeFollowing.resolvers.js
First, it checks if there is a user.

If there is no user, ok :false is set and an error message is sent.

Find the user's information.
take: 5 => Take: 5 information of the people following the user.
  Skip does not skip at first, but there is one skip after that.
last id => The last information seen by the user is called last id and sent to the database.
After the last information, five more follower lists of the user are shown.