Create sendMessage.resolvers.ts
Since you need to use individual private messages, make it a protectedResolver.

User Existence Condition: Checks whether the user exists. -> If the user does not exist return false If the user does exist, create a room.

If the room already exists: If the room already exists, look for a room. If the room ID is not found, return false.

In both cases, a message is created to connect the logged-in user and the other party, and the room ID is also connected.