Update roomUpdates.resolvers.ts
Using withFilter does not receive messages from all rooms.

The first argument of withFilter: It must be a function that returns asyncIterator.

The second argument of withFilter: veriables, where veriables has the id of the room.

Return true and receive a message only when the first argument's roomUpdates.roomId and the second argument's id match.